pdf_llama_chatbot/requirements.txt
# PDF Chatbot: Llama2 + LangChain + RAG & Fine-tuning
# Python 3.10+ recommended

# Core (install main langchain package for chains)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1

# PDF & text
pypdf>=3.0.0
unstructured[pdf]>=0.10.0

# Embeddings & vector store
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.4

# LLM: Hugging Face
transformers>=4.36.0
accelerate>=0.25.0
torch>=2.0.0
torchaudio>=0.9.0   # satisfy torch-audiomentations / torch-pitch-shift; not used by this project

# Fine-tuning (LoRA/QLoRA)
peft>=0.7.0
bitsandbytes>=0.41.0
datasets>=2.14.0
trl>=0.7.0

# Env & CLI
python-dotenv>=1.0.0
